refactor pairing
