+++ Sources/WalletConnectPairing/Types/PairingProtocolMethod.swift
